In these past week and a half, I've been reworking the way I handled the env and export built-ins, by making them linked lists instead of arrays.
Currently the program handles env and export by itself. It handles multiple arguments but it has some issues. The problem the program has now, is that to create the export list,
I use the env list making it so I have to initialize export everytime env changes. I call env_init at the begining of the program and export_init everytime export is called. So if
there are multiple exports called, it keeps building a bigger list lol.

Minor changes to main, still haven't shortened parse_main.

TODO:
For sure fix the export_init conundrum. Maybe by also calling it only at the beginning of the program and everytime something is added to export edit the str (to add quotes and
'declare -x') and sorts the whole export list.

Still don't know if I'll also change cmds to lists.